feat(reranking): integrate MCPReranking configuration into MCPHandler and SearchHandler
