+++ packages/qdrant-loader-mcp-server/src/qdrant_loader_mcp_server/mcp/__init__.py
@@ -1,7 +1,7 @@
 """MCP (Model Context Protocol) implementation for RAG server."""
 
 from .handler import MCPHandler
-from .models import MCPRequest, MCPResponse, MCPReranking
+from .models import MCPRequest, MCPResponse
 from .protocol import MCPProtocol
 
-__all__ = ["MCPProtocol", "MCPHandler", "MCPRequest", "MCPResponse", "MCPReranking"]
+__all__ = ["MCPProtocol", "MCPHandler", "MCPRequest", "MCPResponse"]
