+++ packages/qdrant-loader-mcp-server/src/qdrant_loader_mcp_server/mcp/models.py
@@ -21,16 +21,3 @@ class MCPResponse(BaseModel):
     id: int = Field(..., description="Request ID")
     result: Any | None = Field(None, description="Result of the request")
     error: dict[str, Any] | None = Field(None, description="Error information")
-
-class MCPReranking(BaseModel):
-    """Reranking model."""
-
-    enabled: bool = Field(default=False, description="Enable or disable reranking")
-    model: str = Field(
-        default="cross-encoder/ms-marco-MiniLM-L-12-v2",
-        description="Reranking model to use",
-    )
-    device: str = Field(default="cpu", description="Device to run the reranking model")
-    batch_size: int = Field(
-        default=32, description="Batch size for reranking model inference"
-    )
